Remove Airport's default capacity magic number
Airport initialiser's custom capacity argument now defaults to `null`,
and if no value is provided then `capacity` is set to `defaultCapacity`.
Is this better? Not sure. Maybe it makes it slightly clearer what the
number 20 is doing there, but in fairness it was pretty clear when it
was the default value of the capacity argument.

It means the default capacity property can be accessed in our tests
which is good, but we shouldn't be writing code if the only benefit is
making the testing easier.

So in conclusion, not sure about this.